[feat] #125 Change MyPageNoticeScreen UI
