+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/notice/MyPageNoticeScreen.kt
@@ -1,38 +1,26 @@
-package com.cmc.curtaincall.feature.mypage.announcemnet
+package com.cmc.curtaincall.feature.mypage.notice
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import com.cmc.curtaincall.common.design.R
 import com.cmc.curtaincall.common.design.component.basic.TopAppBarWithBack
+import com.cmc.curtaincall.common.design.component.items.NoticeItem
+import com.cmc.curtaincall.common.design.theme.Cultured
 import com.cmc.curtaincall.common.design.theme.Nero
 import com.cmc.curtaincall.common.design.theme.White
-import com.cmc.curtaincall.common.design.R
-import com.cmc.curtaincall.common.design.extensions.toSp
-import com.cmc.curtaincall.common.design.theme.Bright_Gray
-import com.cmc.curtaincall.common.design.theme.Chinese_Black
-import com.cmc.curtaincall.common.design.theme.Roman_Silver
-import com.cmc.curtaincall.common.design.theme.spoqahansanseeo
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -43,9 +31,6 @@ internal fun MyPageNoticeScreen(
     Scaffold(
         topBar = {
             TopAppBarWithBack(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(54.dp),
                 title = stringResource(R.string.mypage_announcement),
                 containerColor = White,
                 contentColor = Nero,
@@ -73,68 +58,25 @@ private fun MyPageNoticeContent(
         LazyColumn(
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(top = 10.dp)
+                .padding(top = 12.dp)
         ) {
             itemsIndexed(news) { index, item ->
                 NoticeItem(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 20.dp),
+                    modifier = Modifier.fillMaxWidth(),
                     title = "커튼콜 서비스 개편 안내",
                     date = "2023.6.17",
                     isNew = news[index],
+                    notRead = news[index],
                     onClick = onNavigateNoticeDetail
                 )
                 Spacer(
                     modifier = Modifier
                         .fillMaxWidth()
+                        .padding(horizontal = if (news[index]) 0.dp else 20.dp)
                         .height(1.dp)
-                        .background(Bright_Gray)
+                        .background(Cultured)
                 )
             }
         }
-    }
-}
-
-@Composable
-fun NoticeItem(
-    modifier: Modifier = Modifier,
-    title: String,
-    date: String,
-    isNew: Boolean = false,
-    onClick: () -> Unit = {}
-) {
-    Column(
-        modifier = modifier
-            .clickable { onClick() }
-            .padding(vertical = 20.dp)
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Text(
-                text = title,
-                color = Chinese_Black,
-                fontSize = 16.dp.toSp(),
-                fontWeight = FontWeight.Bold,
-                fontFamily = spoqahansanseeo
-            )
-            if (isNew) {
-                Icon(
-                    painter = painterResource(R.drawable.ic_new),
-                    contentDescription = null,
-                    modifier = Modifier
-                        .padding(start = 6.dp)
-                        .size(16.dp),
-                    tint = Color.Unspecified
-                )
-            }
-        }
-        Text(
-            text = date,
-            modifier = Modifier.padding(top = 8.dp),
-            color = Roman_Silver,
-            fontSize = 13.dp.toSp(),
-            fontFamily = spoqahansanseeo,
-            fontWeight = FontWeight.Medium
-        )
     }
 }
